images: adding support for BPF map file name and ifindex
This commit adds a BPF map's name and ifindex to its protobuf image.
ifindex is the index of the network interface to which the BPF map is
attached and can be specified via a parameter while creating the BPF
map (BPF_MAP_CREATE). This commit also provides a default value of
false to the field 'frozen'.

Source files modified:

* images/bpfmap-file.proto

Signed-off-by: Abhishek Vijeev <abhishek.vijeev@gmail.com>